Support GRUB EFI NB for EFI network boot clients.
File system overlay was added as one of the known file systems
in prep-ocsroot.
A better mechanism was used to parse the boot parameters in
ocs-run so that it could deal with that from grub efi network booting.
Bug fixed: set-netboot-1st-efi-nvram failed to keep uEFI network
Bug fixed: ocs-resize-part failed to run resize program for
parted >= 3.