Add ocs_server in the client's boot parameters.
Bug fixed: a workaround to makes sure that client will notify the
lite server so that the next boot the client will enter local booting.
Add ocs_server in the client's boot parameters for Clonezilla lite
server. This is due to a ps issue that somehow it can not list the PID
by this command: ps -C ocs-live-get-img.