Add boot parameter secure_boot_client.
Add boot parameter secure_boot_client so that the mechanism of
secure boot for uEFI client can be enabled with a boot parameters.
The secure boot mechanism for clients is not complete, because there are
two limits for signed shim and grub:
(1) Shim and grub does not honor the proxy offer packet.
    Ref: https://lists.gnu.org/archive/html/grub-devel/2016-04/msg00051.html
         http://lists.thekelleys.org.uk/pipermail/dnsmasq-discuss/2017q1/011347.html
(2) grub lacks the ability to autoload grub2's config file when
netbooting. Ref: https://bugzilla.redhat.com/show_bug.cgi?id=873406