Improve GPT larger disk to smaller one, only -icds
Neglect the line "last-lba:..." for the GPT partition table.
This allows something like:
64 GB disk with a 20 GB NTFS partition, disk clone to >= 20 GB disk.
This is harmless when same sizes of disk cloning or smaller to larger
case.
The option -icds is required when larger disk is cloned/restored to
smaller one, no need to use -k1.
Thanks to panreyes for this idea.
Ref: https://sourceforge.net/p/clonezilla/bugs/342/