Some improvements about ocs-expand-mbr-pt.
No need to check since we are creating new partition table and should
not care about the destination disk's format is GPT or MBR.
Add prompt about the option -icds when failing to creating the partition
table on the smaller disk.
Add "label: dos" in the output config file so that it can be used to
overwrite the GPT disk.