Implement a better mechanism to run ocs-onthefly.
Implement a better mechanism to run ocs-onthefly:
Now ocs-onthefly mainly uses ocs-sr to save the pseudo image,
and let partclone do the device to device clone. This is similar to that
of Clonezilla lite server.
Due to this improvement, some major changes for ocs-othefly in order to
sync with that of ocs-sr:
1. These options have been changed to totally different meanings:
   -d|--destination|--target was -t|--target
   -po|--port was -p|--port
   --net-filter was -i|--filter
   -p|-pa|--postaction was -pa|--postaction
   -u|--use-nuttcp was -u|--use-netcat
2. New options:
   -t|--no-restore-mbr
   -t1|--restore-raw-mbr
   -t2|--no-restore-ebr

By default, the network cloning is changed to use zstd to compress
the data instead of gzip, and the network piping program is changed
to nuttcp from netcat since the latter has too many diverse versions.