ocs-onthefly: revert to use netcat, and more.
Use -np|--net-pipe  PROGRAM instead of -u|--use-netcat
so that it's easier to switch, and it can be
selected in the expert mode.
Revert to use netcat as the default net pipe program since the nuttcp in
Debian/Ubuntu repository is to old and buggy (6.1.2, the latest one now
is 8.2.2).
If a disk has not any partition, it can not be the source for
ocs-onthefly. It will just quit to avoiding confusion.
Add "--rsyncable" for zstd saving.
Since zstdmt is equivalent to zstd -T0, remove "-T${cpu_no}" in
extra_zstdmt_opt.