Improve part to part (local and remote).
Improve part to part (local and remote) for ocs-onthefly.
The previous version did not work for local part to part in interactive
mode, while it works for running in batch command mode. It did not work
for local disk to disk in batch command mode, but works for interactive
mode.
ocs-sr's option -f|--from-part-in-img is changed as -f|--from-part since
it now supports both from an image and a device.