Support an extra dir "root" for root fs with grub.
Support an extra dir "root" for the root fs with grub reinstallation.
This is for the case that Fedora >= 38 uses an extra dir "root"
for root file system, i.e., /root/ under the root partition.
Older system just puts the dirs (boot, usr, var, lib) in the / of root file system.
Thanks to Bob Bobsled for reporting this issue.
Ref: https://sourceforge.net/p/clonezilla/mailman/message/42287717/