ocs-live-repository: added dev=///OCS_LIVE_USB
ocs-live-repository: added dev=///OCS_LIVE_USB so that
the Clonezilla live USB drive can be assigned as the image repo,
especially when Clonezilla live USB drive is booted in "To RAM" mode.
The Clonezilla live USB drive has to be vFAT file system since
it's used to boot in both uEFI and MBR mode. Hence it is not a good
choice for image repo since vFAT file system has many restrictions.
Better to use UUID or LABEL to assign the image repo.