Bug fixed: skip LVM and LUKS info in dev-fs.list.
Bug fixed: removed extra LVM and LUKS info in dev-fs.list.
The issue was introduced on 2024/Aug/1. They should
not be added otherwise it will make clone-over-net of ocs-onthefly
fail because the partition order is messed up. Thanks to kokutou
kiritsugu for reporting this issue and providing an image for testing.
Ref:
https://sourceforge.net/p/clonezilla/discussion/Clonezilla_live/thread/2f352495cb/